Authentication code in a library
