+++ AudioAnalytics/AudioAnalyzer/Program.cs
@@ -9,13 +9,12 @@ using System.Threading;
 using System.Threading.Tasks;
 using Azure.Storage.Blobs;
 using Azure.Storage.Blobs.Models;
+using Common_Authentication;
 using Microsoft.Azure.EventHubs;
 using Microsoft.Azure.EventHubs.Processor;
 using Microsoft.Azure.Management.Media;
 using Microsoft.Azure.Management.Media.Models;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Identity.Client;
-using Microsoft.Rest;
 
 namespace AudioAnalyzer
 {
@@ -24,7 +23,9 @@ namespace AudioAnalyzer
         private const string AudioAnalyzerTransformName = "MyAudioAnalyzerTransformName";
         private const string InputMP4FileName = @"ignite.mp4";
         private const string OutputFolderName = @"Output";
-        private const string TokenType = "Bearer";
+
+        // Set this variable to true if you want to authenticate Interactively through the browser using your Azure user account
+        private const bool UseInteractiveAuth = false;
 
         /// <summary>
         /// The main method of the sample. Please make sure you have set settings in appsettings.json or in the .env file in the root folder
@@ -80,7 +81,7 @@ namespace AudioAnalyzer
             IAzureMediaServicesClient client;
             try
             {
-                client = await CreateMediaServicesClientAsync(config);
+                client = await Authentication.CreateMediaServicesClientAsync(config, UseInteractiveAuth);
             }
             catch (Exception e)
             {
@@ -211,48 +212,6 @@ namespace AudioAnalyzer
             await CleanUpAsync(client, config.ResourceGroup, config.AccountName, AudioAnalyzerTransformName, inputAssetName, outputAssetName, jobName);
         }
 
-        /// <summary>
-        /// Create the ServiceClientCredentials object based on the credentials
-        /// supplied in local configuration file.
-        /// </summary>
-        /// <param name="config">The param is of type ConfigWrapper. This class reads values from local configuration file.</param>
-        /// <returns></returns>
-        // <GetCredentialsAsync>
-        private static async Task<ServiceClientCredentials> GetCredentialsAsync(ConfigWrapper config)
-        {
-            // Use ConfidentialClientApplicationBuilder.AcquireTokenForClient to get a token using a service principal with symmetric key
-
-            var scopes = new[] { config.ArmAadAudience + "/.default" };
-
-            var app = ConfidentialClientApplicationBuilder.Create(config.AadClientId)
-                .WithClientSecret(config.AadSecret)
-                .WithAuthority(AzureCloudInstance.AzurePublic, config.AadTenantId)
-                .Build();
-
-            var authResult = await app.AcquireTokenForClient(scopes)
-                                                     .ExecuteAsync()
-                                                     .ConfigureAwait(false);
-
-            return new TokenCredentials(authResult.AccessToken, TokenType);
-        }
-        // </GetCredentialsAsync>
-
-
-        /// <summary>
-        /// Creates the AzureMediaServicesClient object based on the credentials
-        /// supplied in local configuration file.
-        /// </summary>
-        /// <param name="config">The param is of type ConfigWrapper. This class reads values from local configuration file.</param>
-        /// <returns></returns>
-        private static async Task<IAzureMediaServicesClient> CreateMediaServicesClientAsync(ConfigWrapper config)
-        {
-            var credentials = await GetCredentialsAsync(config);
-
-            return new AzureMediaServicesClient(config.ArmEndpoint, credentials)
-            {
-                SubscriptionId = config.SubscriptionId,
-            };
-        }
 
         /// <summary>
         /// If the specified transform exists, get that transform.
