+++ ContentProtection/OfflineFairPlay/Program.cs
@@ -6,25 +6,28 @@ using Microsoft.Azure.EventHubs.Processor;
 using Microsoft.Azure.Management.Media;
 using Microsoft.Azure.Management.Media.Models;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Identity.Client;
-using Microsoft.Rest;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Common_Authentication;
+
 
 namespace OfflineFairPlay
 {
     class Program
     {
-        private const string TokenType = "Bearer";
         private const string SourceUri = "https://nimbuscdn-nimbuspm.streaming.mediaservices.windows.net/2b533311-b215-4409-80af-529c3e853622/Ignite-short.mp4";
         private static readonly string AdaptiveStreamingTransformName = "MyTransformWithAdaptiveStreamingPreset";
         private static readonly string ContentKeyPolicyName = "FairPlayContentKeyPolicy";
         private static readonly string FairPlayStreamingPolicyName = "FairPlayCustomStreamingPolicyName";
         private static readonly string DefaultStreamingEndpointName = "default";  // Change this to your Endpoint name.
+
+        // Set this variable to true if you want to authenticate Interactively through the browser using your Azure user account
+        private const bool UseInteractiveAuth = false;
+
 
         public static async Task Main(string[] args)
         {
@@ -65,7 +68,7 @@ namespace OfflineFairPlay
             IAzureMediaServicesClient client;
             try
             {
-                client = await CreateMediaServicesClientAsync(config);
+                client = await Authentication.CreateMediaServicesClientAsync(config, UseInteractiveAuth);
             }
             catch (Exception e)
             {
@@ -217,47 +220,6 @@ namespace OfflineFairPlay
             }
         }
 
-        /// <summary>
-        /// Create the ServiceClientCredentials object based on the credentials
-        /// supplied in local configuration file.
-        /// </summary>
-        /// <param name="config">The param is of type ConfigWrapper. This class reads values from local configuration file.</param>
-        /// <returns></returns>
-        // <GetCredentialsAsync>
-        private static async Task<ServiceClientCredentials> GetCredentialsAsync(ConfigWrapper config)
-        {
-            // Use ConfidentialClientApplicationBuilder.AcquireTokenForClient to get a token using a service principal with symmetric key
-
-            var scopes = new[] { config.ArmAadAudience + "/.default" };
-
-            var app = ConfidentialClientApplicationBuilder.Create(config.AadClientId)
-                .WithClientSecret(config.AadSecret)
-                .WithAuthority(AzureCloudInstance.AzurePublic, config.AadTenantId)
-                .Build();
-
-            var authResult = await app.AcquireTokenForClient(scopes)
-                                                     .ExecuteAsync()
-                                                     .ConfigureAwait(false);
-
-            return new TokenCredentials(authResult.AccessToken, TokenType);
-        }
-        // </GetCredentialsAsync>
-
-        /// <summary>
-        /// Creates the AzureMediaServicesClient object based on the credentials
-        /// supplied in local configuration file.
-        /// </summary>
-        /// <param name="config">The param is of type ConfigWrapper. This class reads values from local configuration file.</param>
-        /// <returns></returns>
-        private static async Task<IAzureMediaServicesClient> CreateMediaServicesClientAsync(ConfigWrapper config)
-        {
-            var credentials = await GetCredentialsAsync(config);
-
-            return new AzureMediaServicesClient(config.ArmEndpoint, credentials)
-            {
-                SubscriptionId = config.SubscriptionId,
-            };
-        }
 
         /// <summary>
         /// Create the content key policy that configures how the content key is delivered to end clients 
@@ -294,7 +256,7 @@ namespace OfflineFairPlay
                 Console.WriteLine("Creating a content key policy...");
                 policy = await client.ContentKeyPolicies.CreateOrUpdateAsync(config.ResourceGroup, config.AccountName, contentKeyPolicyName, options);
             }
-            
+
             return policy;
         }
 
@@ -506,7 +468,7 @@ namespace OfflineFairPlay
                 RentalDuration = 0, // in seconds
                 OfflineRentalConfiguration = new ContentKeyPolicyFairPlayOfflineRentalConfiguration()
                 {
-                   StorageDurationSeconds = 300000,
+                    StorageDurationSeconds = 300000,
                     PlaybackDurationSeconds = 500000
                 }
             };
@@ -549,7 +511,7 @@ namespace OfflineFairPlay
                 Console.WriteLine("Creating a Streaming Locator with this name instead: " + locatorName);
             }
 
-            StreamingPolicy customStreamingPolicy = await GetOrCreateCustomStreamingPloliyForFairPlay(client, resourceGroup, accountName, 
+            StreamingPolicy customStreamingPolicy = await GetOrCreateCustomStreamingPloliyForFairPlay(client, resourceGroup, accountName,
                 FairPlayStreamingPolicyName);
 
             Console.WriteLine("Creating a streaming locator...");
@@ -611,7 +573,7 @@ namespace OfflineFairPlay
 
                 streamingPolicy = await client.StreamingPolicies.CreateAsync(resourceGroupName, accountName, streamingPolicyName, streamingPolicy);
             }
-            
+
             return streamingPolicy;
         }
