+++ VideoEncoding/Encoding_PredefinedPreset/Encoding_PredefinedPreset.csproj
@@ -10,16 +10,17 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\Common-Authentication\Common_Authentication.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
     <None Update="appsettings.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-    <None Update="ignite.mp4">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
   </ItemGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <RootNamespace>Encoding_SpriteThumbnail</RootNamespace>
+    <RootNamespace>EncodeHttp</RootNamespace>
     <LangVersion>latest</LangVersion>
   </PropertyGroup>
